#222 split apart views
