+++ pkpdapp/pkpdapp/api/serializers/__init__.py
@@ -6,9 +6,9 @@
 # flake8: noqa F401
 
 
+from .validators import ValidSbml
 from .auce import AuceSerializer
 from .biomarker_type import BiomarkerTypeSerializer
-from .dataset import DatasetSerializer
 from .dose import DoseSerializer
 from .models import (
     PharmacokineticSerializer,
@@ -17,10 +17,10 @@ from .models import (
     PkpdSerializer
 )
 from .nca import NcaSerializer
-from .project import ProjectSerializer
+from .project import ProjectSerializer, ProjectAccessSerializer
 from .protocol import ProtocolSerializer
+from .dataset import DatasetSerializer
 from .subject import SubjectSerializer
 from .unit import UnitSerializer
 from .user import UserSerializer
-from .validators import ValidSbml
 from .variables import VariableSerializer, StoredVariableSerializer
